Created a config.py file where the base field and its variables are stored, to be accessed by all files.
This is not the ideal solution because this cannot be changed directly from main. If the base_field variables are defined in main, however, they will not be defined in any of the other files unless they import main (which is bad).

Fix a bunch of imports to remove circularity. In particular, most "from A import B" have been replace by "import A" and then "A.B" in the code, which creates a lot of clutter. Not ideal...