Implemented SAGE import
A SAGE interpreter must be set in the IDE (it's not trivial, environment variables must be set, paths must be given and a couple of errors in the console must be troubleshoted). The main file will run and the base_field and its variables will be passed into the internal files (via the q_matrix). The variables are not passed explicitly and hence I removed the variables attribute from FreeAlgebra and PBWAlgebra (and set universe).

Ran a "Reformat code" from PyCharm to uniformize syntax.

To avoid circular imports, made a local import of element.c_bilinear in letters.py and removed the local import